Configure journal on a platform-specific basis
We can only build/link with sd_journal on linux systems. On other
systems, just configure and build a dummy impl.
