+++ collector/logs/sources/journal/journal_linux_test.go
